Desktop/solana bot/requirements.txt
python-telegram-bot
flask
requests
python-dotenv
gunicorn  <-- ADD THIS LINE